requirements.txt
tensorflow
Flask
flask-cors
Pillow
numpy
gunicorn